--- src/components/form-submit/examples/form-submit-example-basic.js
@@ -1,21 +1,0 @@
-/*
-Basic.
-*/
-import React from 'react';
-import FormSubmit from '../form-submit';
-
-export default class Example extends React.Component {
-  render() {
-    return (
-      <FormSubmit
-        onSubmit={
-          (/* e */) => {
-            /* console.log(e); */
-          }
-        }
-        id="submit"
-        label="Submit me"
-      />
-    );
-  }
-}
