+++ src/components/form-submit/examples/form-submit-example-basic.tsx
@@ -1,0 +1,19 @@
+/*
+Basic.
+*/
+import React from 'react';
+import FormSubmit from '../form-submit';
+
+export default function Example() {
+
+  const onSubmit = (e) => {
+    console.log(e);
+  };
+
+  return (
+    <FormSubmit
+      onSubmit={onSubmit}
+      label="Submit me"
+    />
+  );
+}
